+++ src/buildTranspilerFile.js
@@ -3,6 +3,7 @@ const path = require("path").posix;
 
 const getModuleFilesOfType = require('./getModuleFilesOfType.js');
 const relativizeCssUrls = require('./relativizeCssUrls.js');
+const requireAdaptor = require('./requireAdaptor.js');
 
 async function buildTranspilerFile(settings) {
 	const assetList = await Promise.all(Object.keys(settings.collations).map(collationName => {
@@ -22,7 +23,8 @@ async function each(settings, name, modules) {
 		isDev,
 	} = settings;
 	
-	let types = cssTranspilerAdaptor.inputTypes;
+	let transpilerAdaptor = requireAdaptor(base, cssTranspilerAdaptor)
+	let types = transpilerAdaptor.inputTypes;
 	let buildDir = isDev ? buildDevDir : buildDistDir;
 	let buildCssFile = `${name}.css`;
 	let entryFile = `${buildCssFile}.${types[0]}`;
@@ -56,7 +58,7 @@ async function each(settings, name, modules) {
 	await fs.promises.mkdir(path.dirname(entryPath), {recursive: true});
 	await fs.promises.mkdir(path.dirname(outputPath), {recursive: true});
 
-	cssTranspilerAdaptor.run({sourcePaths, entryPath, outputPath, isDev});
+	transpilerAdaptor.run({sourcePaths, entryPath, outputPath, isDev});
 	
 	if (!isDev) {
 		let tempDirToDelete = path.resolve(path.join(base, buildDir, buildTempDir));